Move delete button to entity property columns control SN-640
