+++ src/Saritasa.NetForge.Blazor/Controls/EntityPropertyColumns.razor
@@ -104,3 +104,15 @@
         </TemplateColumn>
     }
 }
+
+@if (!IsNavigationEntity)
+{
+    <TemplateColumn T="object">
+        <CellTemplate>
+            <MudButton Size="@Size.Small" Variant="Variant.Text" Color="Color.Secondary"
+                       OnClick="@(() => ShowDeleteEntityConfirmationAsync(context.Item))">
+                Delete
+            </MudButton>
+        </CellTemplate>
+    </TemplateColumn>
+}
