+++ src/Saritasa.NetForge.Blazor/Controls/NavigationDialog.razor
@@ -17,7 +17,7 @@
 
             <MudDataGrid Items="collection">
                 <Columns>
-                    <EntityPropertyColumns Properties="Properties" />
+                    <EntityPropertyColumns Properties="Properties" IsNavigationEntity="true" />
                 </Columns>
             </MudDataGrid>
         }
