+++ src/Saritasa.NetForge.Blazor/Pages/EntityDetails.razor.cs
@@ -70,28 +70,6 @@ public partial class EntityDetails : MvvmComponentBase<EntityDetailsViewModel>
         NavigationService.NavigateTo<CreateEntityViewModel>(parameters: StringId);
     }
 
-    private async void ShowDeleteEntityConfirmationAsync(object source)
-    {
-        var parameters = new DialogParameters();
-        parameters.Add(nameof(ConfirmationDialog.ContentText), "Are you sure you want to delete this record?");
-        parameters.Add(nameof(ConfirmationDialog.ButtonText), "Yes");
-        parameters.Add(nameof(ConfirmationDialog.Color), Color.Error);
-
-        var result = await (await DialogService.ShowAsync<ConfirmationDialog>("Delete", parameters)).Result;
-        if (!result.Canceled)
-        {
-            try
-            {
-                await ViewModel.DeleteEntityAsync(source, CancellationToken.None);
-            }
-            catch (Exception ex)
-            {
-                Snackbar.Add($"Failed to delete record due to error: {ex.Message}", Severity.Error);
-                Logger.LogError("Failed to delete record due to error: {ex.Message}", ex.Message);
-            }
-        }
-    }
-
     private void NavigateToEditing(DataGridRowClickEventArgs<object> row)
     {
         var primaryKeyValues = ViewModel.Model.Properties
