+++ src/Saritasa.NetForge.Mvvm/ViewModels/EntityDetails/EntityDetailsViewModel.cs
@@ -171,13 +171,4 @@ public class EntityDetailsViewModel : BaseViewModel
     {
         DataGrid?.ReloadServerData();
     }
-
-    /// <summary>
-    /// Delete entity.
-    /// </summary>
-    public async Task DeleteEntityAsync(object entity, CancellationToken cancellationToken)
-    {
-        await entityService.DeleteEntityAsync(entity, entity.GetType(), cancellationToken);
-        DataGrid?.ReloadServerData();
-    }
 }
